requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0 
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.2.11
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
msgpack==1.0.8
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.1.1
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stripe==9.10.0
twilio==9.0.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
django-cors-headers==4.4.0